+++ tests/valid/boolean_operators/multi_short_circuit.rs
@@ -3,8 +3,8 @@ return: 3
 skip: 1
 */
 
-main(): int {
-    a: int = 0;
+main(): i32 {
+    a: i64 = 0;
     a || (a = 3) || (a = 4);
     return a;
 }
