+++ tests/valid/conditional_expression/multiple_ternary.rs
@@ -2,8 +2,8 @@
 return: 10
 */
 
-main(): int {
-    a: int = 1 > 2 ? 3 : 4;
-    b: int = 1 > 2 ? 5 : 6;
+main(): i32 {
+    a: i64 = 1 > 2 ? 3 : 4;
+    b: i64 = 1 > 2 ? 5 : 6;
     return a + b;
 }
