+++ tests/valid/functions/multi_arg.rs
@@ -2,10 +2,10 @@
 return: 4
 */
 
-sub_3(x: int, y: int, z: int): int {
+sub_3(x: i64, y: i64, z: i64): i64 {
     return x - y - z;
 }
 
-main(): int {
+main(): i32 {
     return sub_3(10, 4, 2);
 }
