+++ tests/valid/loops/for_nested_scope.rs
@@ -2,16 +2,16 @@
 return: 3
 */
 
-main(): int {
-    i: int = 0;
-    j: int = 0;
+main(): i32 {
+    i: i64 = 0;
+    j: i64 = 0;
 
-    for (i: int = 100; i > 0; i = i - 1) {
-        i: int = 0;
-        j: int = j * 2 + i;
+    for (i: i64 = 100; i > 0; i = i - 1) {
+        i: i64 = 0;
+        j: i64 = j * 2 + i;
     }
 
-    k: int = 3;
+    k: i64 = 3;
 
     return j + k;
 }
