+++ tests/valid/loops/for_variable_shadow.rs
@@ -2,12 +2,12 @@
 return: 65
 */
 
-main(): int {
-    i: int = 0;
-    j: int = 0;
+main(): i32 {
+    i: i64 = 0;
+    j: i64 = 0;
     for (i = 0; i < 10; i = i + 1) {
-        k: int = i;
-        for (i: int = k; i < 10; i = i + 1)
+        k: i64 = i;
+        for (i: i64 = k; i < 10; i = i + 1)
             j = j + 1;
     }
     return j + i;
