+++ tests/valid/loops/skip_empty_post.rs
@@ -2,9 +2,9 @@
 return: 30
 */
 
-main(): int {
-    sum: int = 0;
-    for (i: int = 0; i < 10;) {
+main(): i32 {
+    sum: i64 = 0;
+    for (i: i64 = 0; i < 10;) {
         i = i + 1;
         // this is the modulo operator, which is not supported yet
         if (i - (i / 2) * 2)
