+++ tests/valid/scoping/consecutive_blocks.rs
@@ -2,10 +2,10 @@
 return: 1
 */
 
-main(): int {
-    a: int = 1;
+main(): i32 {
+    a: i64 = 1;
     {
-        a: int = 2;
+        a: i64 = 2;
     }
     {
         return a;
