+++ tests/valid/scoping/nested_if.rs
@@ -2,13 +2,13 @@
 return: 4
 */
 
-main(): int {
-    a: int = 0;
+main(): i32 {
+    a: i64 = 0;
     if (a) {
-        b: int = 2;
+        b: i64 = 2;
         return b;
     } else {
-        c: int = 3;
+        c: i64 = 3;
         if (a < c) {
             return 4;
         } else {
